+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * updated Lets Encrypt Certbot to version 0.10.1
 * updated nginx to version 1.11.8
 * changed from "dotdeb" to "sury" repository for PHP and PHP-FPM
-* updated PHP and PHP-FPM to version 7.1.1
+* updated PHP and PHP-FPM to version 7.1.2
 * updated Composer to version 1.3.1
 * updated NodeJS to version 6.9.4
 * updated NPM to version 4.1.1