+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * updated Debian Jessie to version 8.7
 * updated VirtualBox Guest additions to version 5.1.14
 * updated Lets Encrypt Certbot to version 0.11.1
+* updated PostgreSQL to version 9.6.2
 * updated nginx to version 1.11.8
 * changed from "dotdeb" to "sury" repository for PHP and PHP-FPM
 * updated PHP and PHP-FPM to version 7.1.2