+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * updated nginx to version 1.11.8
 * changed from "dotdeb" to "sury" repository for PHP and PHP-FPM
 * updated PHP and PHP-FPM to version 7.1.2
-* updated Composer to version 1.3.1
+* updated Composer to version 1.3.2
 * updated NodeJS to version 6.9.4
 * updated NPM to version 4.1.1
 * updated Newman to version 3.4.2